Feature/remove catkin simple (#8)
* restructure to conform better to ros standard layout and fix compile warnings

* minor clean up to python cmake

* fix warnings in object db and standardize message dependencies better

+++ kimera_scene_graph/include/kimera_scene_graph/dynamic_scene_graph.h
@@ -266,7 +266,7 @@ class DynamicSceneGraph {
   double pgo_trans_threshold_ = 0.05;
   double pgo_rot_threshold_ = 0.005;
   double beta_diff_threshold_ = 0.1;
-  int prune_threshold_ = 0;
+  size_t prune_threshold_ = 0;
   bool single_sequence_smpl_mode_ = false;
   bool draw_skeleton_edges_ = false;
   std::string human_topic_{"human_meshes"};
